March 19, 2007
The starter package makes it easier for first-time users to download FR3D and start using it.  A new option has been added to the Display Candidates menu, the ability to make scatterplots of the parameters describing pairs of nucleotides in the candidates.  Known bugs:  the alignment of candidates is sometimes incorrect; figure windows sometimes have incorrect names.